requirements.txt
alembic==1.10.4
aniso8601==9.0.1
APScheduler==3.10.1
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
environs==9.5.0
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.28.2
rq==1.13.0
six==1.16.0
SQLAlchemy==1.4.39
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
Werkzeug==2.2.3